pyproject.toml
[project]
name = "spike-ai-hackathon"
version = "1.0.0"
description = "Production-ready multi-agent AI backend for Spike AI BuildX Hackathon"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "google-analytics-data>=0.18.0",
    "google-auth>=2.25.0",
    "gspread>=5.12.0",
    "litellm>=1.17.0",
    "tenacity>=8.2.3",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
